requirements.txt
opentelemetry-api
opentelemetry-sdk
flask
opentelemetry-exporter-otlp
prometheus-client
